GREAT

+++ srcs/vector/vectors_op3.c
@@ -22,7 +22,7 @@ t_vector	v_double_div(t_vector *a, double b)
 	return (r);
 }
 
-int			max(int a, int b)
+double		max(double a, double b)
 {
 	return (a > b ? a : b);
 }
